finance_tracker/requirements.txt
streamlit==1.40.0
supabase==2.10.0
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.5.0
prophet==1.1.5
plotly==5.24.0
matplotlib==3.9.0
seaborn==0.13.0
pytest==8.3.0
python-dateutil==2.9.0
# scipy==1.11.0
joblib==1.3.2
